Force explicit listing of header files in the CMake.
This new approach may be a bit annoying (since it re-runs CMake each
time a header is modified), but this about the only way I have found to
ensure that all headers are explicitly listed in the CMake project.

It is important to list project headers for proper integration into IDEs
such as CLion, Visual Studio or Xcode. The current implementation does
not yet define source_group() or FOLDER names for the different targets,
but this is a start.